% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saleOrderTable.R
\name{saleOrderTable_customer_query}
\alias{saleOrderTable_customer_query}
\title{嘉好客户往来对账单v1}
\usage{
saleOrderTable_customer_query(token, FCustName)
}
\arguments{
\item{token}{中台数据库token}

\item{FNumber}{单据编号}

\item{FName}{ERP账套名}
}
\value{

}
\description{
嘉好客户往来对账单v1
}
\examples{
saleOrderTable_customer_query('C0426D23-1927-4314-8736-A74B2EF7A039','ZANHANG S.A.')
}
