% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saleOrderTable.R
\name{saleOrderTable_date_organization_query}
\alias{saleOrderTable_date_organization_query}
\title{嘉好对账单-日期组织}
\usage{
saleOrderTable_date_organization_query(
  token,
  FCustName,
  FStartDate,
  FEndDate,
  FOrgNumber
)
}
\arguments{
\item{token}{中台数据库token}

\item{FNumber}{单据编号}

\item{FName}{ERP账套名}
}
\value{

}
\description{
嘉好对账单-日期组织
}
\examples{
saleOrderTable_date_organization_query('C0426D23-1927-4314-8736-A74B2EF7A039','ZANHANG S.A.','2019-01-01','2023-08-22','102')
}
