% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saleOrderTable.R
\name{saleOrderTable_day_query}
\alias{saleOrderTable_day_query}
\title{嘉好按天合并对账单v2}
\usage{
saleOrderTable_day_query(token, FCustName)
}
\arguments{
\item{token}{中台数据库token}

\item{FNumber}{单据编号}

\item{FName}{ERP账套名}
}
\value{

}
\description{
嘉好按天合并对账单v2
}
\examples{
saleOrderTable_day_query('C0426D23-1927-4314-8736-A74B2EF7A039','ZANHANG S.A.')
}
