% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saleOrderTable.R
\name{saleOrderTable_note_query}
\alias{saleOrderTable_note_query}
\title{嘉好对账单应收备注}
\usage{
saleOrderTable_note_query(token, FBillNO, RemarkText)
}
\arguments{
\item{token}{中台数据库token}

\item{FNumber}{单据编号}

\item{FName}{ERP账套名}
}
\value{

}
\description{
嘉好对账单应收备注
}
\examples{
saleOrderTable_note_query('C0426D23-1927-4314-8736-A74B2EF7A039','AR00047594','')
}
