% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saleOrderTable.R
\name{saleOrderTable_research_query}
\alias{saleOrderTable_research_query}
\title{嘉好对账单-研发订单}
\usage{
saleOrderTable_research_query(token, FBillNO, FRdNumber)
}
\arguments{
\item{token}{中台数据库token}

\item{FNumber}{单据编号}

\item{FName}{ERP账套名}
}
\value{

}
\description{
嘉好对账单-研发订单
}
\examples{
saleOrderTable_research_query('C0426D23-1927-4314-8736-A74B2EF7A039','XSDD-100-230818-001','')
}
